Move `Assume` to spring-boot-test-support
Fixes gh-10866
